Fix writing to the output fd
The previous implementation (partially stolen from the syscall_intercept
example code) was trying to be to clever with atomics and buffering and
was broken as a result. The "fix" is to not attempt to do any buffering
in the plugin whatsoever and just issue a write system call. The OS
should buffer anyway.